% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deductive-pkg.R
\docType{package}
\name{deductive}
\alias{deductive}
\title{Deductive Data Correction and Imputation}
\description{
Use data validation restrictions to estimate missing values
or trace and repair certain errors.
}
